+++ include/Chassis2D/Particle.h
@@ -26,7 +26,7 @@ class Particle : public Object{
   int GetTime() {return time;}
   int GetLifetime() { return lifetime; }
 
-  int Create(App *a, int x, int y);
+  int Create(App *a, float x, float y);
   void SetLifetime(int set) {lifetime = set;}
   void SetTime(int set) {time = set;}
